Fix access violation when using multiple async clients.
The GnuTls wrapper always freed the dynamic library when deinit was called, even if init was called multiple times.
Add a use count to keep track of the number of active users and only free the library if none are left.
Add a lock to prevent race conditions when loading and freeing the library.